// frontend/src/app/core/enums/type-alerte.enum.ts
export enum TypeAlerte {
    STOCK_FAIBLE = 'STOCK_FAIBLE',
    DATE_EXPIRATION = 'DATE_EXPIRATION',
    COMMANDE_RETARD = 'COMMANDE_RETARD',
    COMMANDE_ANNULEE = 'COMMANDE_ANNULEE',
    PRODUCTION_RETARD = 'PRODUCTION_RETARD',
    PRODUCTION_ANNULEE = 'PRODUCTION_ANNULEE',
    VENTE_EXCEPTIONNELLE = 'VENTE_EXCEPTIONNELLE',
    VENTE_ANNULEE = 'VENTE_ANNULEE',
    PERTE_STOCK = 'PERTE_STOCK',
    ALERTE_SECURITE = 'ALERTE_SECURITE',
    PRODUCTION_TERMINEE = 'PRODUCTION_TERMINEE',
    PROBLEME_SIGNALE = 'PROBLEME_SIGNALE',
    NOUVELLE_COMMANDE = 'NOUVELLE_COMMANDE'
}